0.13 beta2 (#65)
* Issue: Column delegate (combobox) drawn "without parent" (WIP)

* FIX: Propery set parent of EnumerationComboBox

- From within the EnumerationItemDelegate

* Code analysis refactoring

* Remove dynamic casting

* Remove default left-alignment call

* FIX: Proper usage of success vs ok (pointer)

* Code analysis refactoring

* Refactor

* File suffix -> extension

* File suffix -> extension 2

* Adjust pre-allocated memory count to new number of aircraft types

* Add locations

* Add locations

* Added locations

* Fix linebreaks

* Also update info group when updating a location

* Validate CSV migration file

- Change order of CSV migration file columns (in analogy how they are
exported as Sky Dolly CSV location file)

* Additional locations

+++ src/Persistence/src/Dao/SQLite/SQLitePrimaryFlightControlDao.cpp
@@ -119,9 +119,9 @@ std::vector<PrimaryFlightControlData> SQLitePrimaryFlightControlDao::getByAircra
             PrimaryFlightControlData data;
 
             data.timestamp = query.value(timestampIdx).toLongLong();
-            data.rudderPosition = query.value(rudderPositionIdx).toDouble();
-            data.elevatorPosition = query.value(elevatorPositionIdx).toDouble();
-            data.aileronPosition = query.value(aileronPositionIdx).toDouble();
+            data.rudderPosition = query.value(rudderPositionIdx).toInt();
+            data.elevatorPosition = query.value(elevatorPositionIdx).toInt();
+            data.aileronPosition = query.value(aileronPositionIdx).toInt();
 
             primaryFlightControlData.push_back(std::move(data));
         }
